Update UI styles in App.svelte
In this commit, some styles were updated in the App.svelte file. These changes include setting the overflow property in the .chatContainer class to hidden, updating the .navbar-brand class to add some appearance enhancements, and changing the color as well as the content of the <a> element within the navbar.