Rename thread page and improve thread display
The 'Threads.svelte' page has been renamed to 'ThreadsList.svelte'. Additionally, some enhancements have been made to thread display, including the calculation of time from the 'updatedAt' timestamp and the use of 'fromNow' for the display of time. A split function has been added to the dropdown to duplicate threads and instructions separately.